requirements.txt
django==3.1.7
djangorestframework==3.12.2
django-video-encoding==1.0.0
django-rest-swagger==2.2.0
psycopg2-binary==2.8.6